+++ src/main/java/xyz/nikitacartes/easyauth/commands/AccountCommand.java
@@ -101,7 +101,11 @@ public class AccountCommand {
 
         // Different thread to avoid lag spikes
         THREADPOOL.submit(() -> {
+            //? if >= 1.20.3 {
             String username = player.getNameForScoreboard();
+            //?} else {
+            /*String username = player.getName().getString();
+            *///?}
             if (AuthHelper.checkPassword(playerAuth, pass.toCharArray()) == AuthHelper.PasswordOptions.CORRECT) {
                 DB.deleteUserData(username);
                 langConfig.accountDeleted.send(source);
@@ -162,8 +166,13 @@ public class AccountCommand {
 
         THREADPOOL.submit(() -> {
             if (AuthHelper.checkPassword(playerAuth, password.toCharArray()) == AuthHelper.PasswordOptions.CORRECT) {
+                //? if >= 1.20.3 {
+                String username = player.getNameForScoreboard();
+                //?} else {
+                /*String username = player.getName().getString();
+                *///?}
                 try {
-                    if (!isValidUsername(player.getNameForScoreboard())) {
+                    if (!isValidUsername(username)) {
                         langConfig.accountNotFound.send(source);
                         return;
                     }