+++ src/main/java/xyz/nikitacartes/easyauth/integrations/VanishIntegration.java
@@ -1,7 +1,9 @@
 package xyz.nikitacartes.easyauth.integrations;
 
 import me.drex.vanish.api.VanishAPI;
+//? if != 1.20.2 {
 import me.drex.vanish.util.VanishedEntity;
+//?}
 import net.minecraft.server.network.ServerPlayerEntity;
 
 public class VanishIntegration {
@@ -11,6 +13,8 @@ public class VanishIntegration {
 
     public static void setVanished(ServerPlayerEntity player, boolean vanished) {
         VanishAPI.setVanish(player, vanished);
+        //? if != 1.20.2 {
         ((VanishedEntity) player).vanish$setDirty();
+        //?}
     }
 }