+++ src/main/java/xyz/nikitacartes/easyauth/mixin/ServerLoginNetworkHandlerMixin.java
@@ -50,7 +50,13 @@ public abstract class ServerLoginNetworkHandlerMixin {
             method = "onHello(Lnet/minecraft/network/packet/c2s/login/LoginHelloC2SPacket;)V",
             at = @At(
                     value = "INVOKE",
+                    //? if >= 1.20.2 {
                     target = "Lnet/minecraft/server/MinecraftServer;isOnlineMode()Z"
+                    //?} else {
+                    /*target = "Lcom/mojang/authlib/GameProfile;<init>(Ljava/util/UUID;Ljava/lang/String;)V",
+                    shift = At.Shift.AFTER,
+                    remap = false
+                    *///?}
             ),
             cancellable = true
     )
@@ -65,9 +71,15 @@ public abstract class ServerLoginNetworkHandlerMixin {
 
                 if (playerData.onlineAccount == PlayerEntryV1.OnlineAccount.FALSE) {
                     LogDebug("Player " + username + " is forced to be offline");
+                    //? if >= 1.20.2 {
                     state = ServerLoginNetworkHandler.State.VERIFYING;
 
                     this.profile = new GameProfile(Uuids.getOfflinePlayerUuid(packet.name()), packet.name());
+                    //?} else {
+                    /*state = ServerLoginNetworkHandler.State.READY_TO_ACCEPT;
+
+                    this.profile = new GameProfile(null, packet.name());
+                    *///?}
                     ci.cancel();
                     return;
                 }
@@ -78,11 +90,19 @@ public abstract class ServerLoginNetworkHandlerMixin {
                 if (!matcher.matches()) {
                     // Player definitely doesn't have a mojang account
                     LogDebug("Player " + username + " doesn't have a valid username for Mojang account");
+                    //? if >= 1.20.2 {
                     state = ServerLoginNetworkHandler.State.VERIFYING;
+                    //?} else {
+                    /*state = ServerLoginNetworkHandler.State.READY_TO_ACCEPT;
+                    *///?}
                     playerData.onlineAccount = PlayerEntryV1.OnlineAccount.FALSE;
                     playerData.update();
 
+                    //? if >= 1.20.2 {
                     this.profile = new GameProfile(Uuids.getOfflinePlayerUuid(packet.name()), packet.name());
+                    //?} else {
+                    /*this.profile = new GameProfile(null, packet.name());
+                    *///?}
                     ci.cancel();
                 } else {
                     if (isValidUsername(username)) {
@@ -91,12 +111,20 @@ public abstract class ServerLoginNetworkHandlerMixin {
                         playerData.update();
                         // Authentication continues in the original method
                     } else {
+                        //? if >= 1.20.2 {
                         state = ServerLoginNetworkHandler.State.VERIFYING;
+                        //?} else {
+                        /*state = ServerLoginNetworkHandler.State.READY_TO_ACCEPT;
+                        *///?}
 
                         playerData.onlineAccount = PlayerEntryV1.OnlineAccount.FALSE;
                         playerData.update();
 
+                        //? if >= 1.20.2 {
                         this.profile = new GameProfile(Uuids.getOfflinePlayerUuid(packet.name()), packet.name());
+                        //?} else {
+                        /*this.profile = new GameProfile(null, packet.name());
+                        *///?}
                         ci.cancel();
                     }
                 }