+++ src/main/java/xyz/nikitacartes/easyauth/mixin/ServerPlayNetworkHandlerMixin.java
@@ -29,8 +29,10 @@ public abstract class ServerPlayNetworkHandlerMixin {
                     value = "INVOKE",
                     //? if >= 1.20.5 {
                     target = "Lnet/minecraft/server/network/ServerPlayNetworkHandler;validateAcknowledgment(Lnet/minecraft/network/message/LastSeenMessageList$Acknowledgment;)Ljava/util/Optional;",
+                    //?} else if >= 1.20.3 {
+                    //target = "Lnet/minecraft/server/network/ServerPlayNetworkHandler;validateMessage(Lnet/minecraft/network/message/LastSeenMessageList$Acknowledgment;)Ljava/util/Optional;",
                     //?} else {
-                    /*target = "Lnet/minecraft/server/network/ServerPlayNetworkHandler;validateMessage(Lnet/minecraft/network/message/LastSeenMessageList$Acknowledgment;)Ljava/util/Optional;",
+                    /*target = "Lnet/minecraft/server/network/ServerPlayNetworkHandler;validateMessage(Ljava/lang/String;Ljava/time/Instant;Lnet/minecraft/network/message/LastSeenMessageList$Acknowledgment;)Ljava/util/Optional;",
                     *///?}
                     shift = At.Shift.BEFORE
             ),