+++ src/main/java/xyz/nikitacartes/easyauth/mixin/WorldSaveHandlerMixin.java
@@ -5,7 +5,9 @@ import com.llamalad7.mixinextras.sugar.Local;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtIo;
+//? if >= 1.20.3 {
 import net.minecraft.nbt.NbtSizeTracker;
+//?}
 import net.minecraft.util.Uuids;
 import net.minecraft.world.WorldSaveHandler;
 import org.spongepowered.asm.mixin.Final;
@@ -72,7 +74,11 @@ public class WorldSaveHandlerMixin {
                 /^File file = new File(this.playerDataDir, Uuids.getOfflinePlayerUuid(player.getGameProfile().getName()) + ".dat");
                 if (file.exists() && file.isFile()) {
                     try {
+                        //? if >= 1.20.3 {
                         cir.setReturnValue(NbtIo.readCompressed(file.toPath(), NbtSizeTracker.ofUnlimitedBytes()));
+                        //?} else {
+                        /^¹cir.setReturnValue(NbtIo.readCompressed(file));
+                        ¹^///?}
                     } catch (IOException e) {
                         LogWarn(String.format("Failed to load player data for: %s", playername));
                     }