+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ----
 ### 3.3.4
-##### Minecraft 1.20.3-1.20.4, 1.20.5-1.20.6, 1.21-1.21.1, 1.21.2-1.21.4, 1.21.5, 1.21.6
+##### Minecraft 1.20[.X], 1.21[.X]
 #### Fix
 - Re-register `register` command if `single-use-global-password` was changed